Sửa lỗi tổng đơn hàng

+++ Avengers/Avengers/db/migrate/20180618155231_create_orders.rb
@@ -6,7 +6,7 @@ class CreateOrders < ActiveRecord::Migration[5.2]
       t.string :email
       t.decimal :sdt
       t.text :ghichu
-      t.decimal :total
+      t.integer :total
 
       t.timestamps
     end
